Include Bundlor-generated manifests in jars
In the 2.0.0.M5 release bundlor successfully created manifests at
target/classes/META-INF/MANIFEST.MF within each subproject directory,
however they were not actually picked up by the maven-jar-plugin when
creating each archive. The samples parent pom has now been configured
to explictly specify the 'manifestFile' and refer to the bundlor-generated
manifest.

Additionally, the spring-integration-samples/pom.xml was updated with
a similar configuration for the maven-jar-plugin, but in this case the
spring-integration-parent/pom.xml configuration is being overwritten.
The samples projects do not use Bundlor, and thus there is no generated
manifest in the location mentioned above. This causes the maven-jar-plugin
to error out for the samples, and as a workaround there is now a zero-length
file at spring-integration-samples/src/main/resources/META-INF/MANIFEST.MF
that samples refer to during jar/war packaging. Ugly, but effective.