Add ability to generate poms for use at build time
Generate poms from Gradle metadata with

    `gradle generatePom`

This will create a root pom with a <modules> section as well as a pom
for every individual module.

These poms are suitable for use with m2eclipse (e.g., File->Import->
Existing Maven project), or at the command line for basic build goals
such as `mvn test`.

These poms are not capable of producing distribution artifacts or
deployment of artifacts.

pom.xml and target will remain in .gitignore as these artifacts are
transient and for developer convenience only.