Supply ssh private key during docs upload
Prior to this change, a graphical Ivy-branded authentication dialog was
appearing when attepmting to scp the docs zip file to
static.springframework.org.  This was because the 'keyFile' property
of the uploadArchives task was not being set.  It is being set correctly
now, based on the value of the 'sshPrivateKey' project property that
must be set in gradle.properties.  This value has been set on the
build server's gradle.properties file, such that the INT-NIGHTLY build
can succeed in pushing docs during the build.