DefaultAmqpHeaderMapper refactoring
  INT-2016, INT-2017 transient headers are no longer mapped
  INT-2090 prevent overwriting of existing headers

Most of INT-2016 is already covered by the other changes made for INT-2017.
However, it was mentioned in INT-2016 that the JsonMessageConverter was
setting a __TypeId__ header, and that header was then being overwritten
if one was present in the MessageHeaders. This change now prevents any
header value from being overwritten so that the MessageConverter or
any process that explicitly added headers to the Message would take
precedence over the HeaderMapper which may write headers subsequently.

excluding id, timestamp, error/reply channel only

updated comment issue link