INT-2434 v2 approach with two FactoryBeans This approach is radically different from the previous one. It has no Factory Post Processor but instead it has two FactoryBeans for Input and Output channels.
The InputChannelNameFactoryBean is constructed with the channel name. This value (placeholder or expression) is
resolved naturally and during the getObject() call the check is made if channel exists and if not it is created
and the channel name is returned

The OutputChannelFactoryBean is constructed with the channel name. This value (placeholder or expression) is
resolved naturally and during the getObject() call it calls:

 this.beanFactory.getBeansOfType(InputChannelNameFactoryBean.class);

which forces all InputChannelNameFactoryBeans to be created, thus creating all the input-channels that are not explicitly defined.
Than it simply returns already created channel.

Pay attention to the updated test especially 'serviceA' where output-channel is an expression. The 'serviceB' input-channel is the same expression and the trick is that 'serviceB' input-channel must be resolved first so the actual channel is created and can be resolved as an output-channel of 'serviceA'