INT-2636 Gateway Timeout Changes
Previously, the timeouts in @Gateway method annotations always
overrode the default timeout if specified in, say, XML.

Now, the @Gateway timeouts are applied only if

* The equivalent default was not set
* The annotation value is something other than the default (Long.MIN_VALUE)

Note: This is a breaking change for one scenario:

<int:gateway ... default-reply-timeout=1000 ... />

where the interface has a method annotated with @Gateway(replyTimeout=Long.MIN_VALUE)

In other words, when the user wants a default timeout on the gateway
but wants to override it for just the annotated method and happened to use
the new default value.

The @Gateway annotation does not allow us to detect whether its
attribute value came from the default or was explicitly set.

The work-around for this breaking change is to use a different negative
number to configure an infinite timeout:

@Gateway(replyTimeout=-1)