INT-2733/4/6 Configure Redis Serializers
Add support for serializers to the RedisStore Inbound adapters.

Fix the serializer support for RedisStore outbound adapters
to ensure that serializers cannot be set if adapter is
bootstrapped with a RedisTemplate.

Add support for serializers to Redis Inbound/Outbound adapters.
NOTE: unlike the RedisStore adapters Redis in/out adapters only
use one serializer, so the name of the attribute
is 'serializer' and, on the Redis Outbound adapter,
it corresponds to the RedisTemplate.valueSerializer.

INT-2733/4/6

Fix RedisCollectionPopulatingMessageHandler to
ensure that the checks for setting serializers also exist
at the class level in case an instance is created outside
of the namespace parser's control.

INT-2733/4/6 polishing

INT-2733/4/6 polishing

INT-2733/4/6 polishing

INT-2733/4/6 Polishing