INT-2844 Fix Http Test If Building With Spring 3.0
In Spring 3.1, DefaultResponseErrorHandler.getResponseBody() is
tolerant of a null InputStream on the response. With Spring 3.0
a null response causes an assertion exception.

The mock in the tests was not set up to return an InputStream
and so the test failed when building with Spring 3.0.

Configure the mock to return an empty InputStream.

Also added a text response to response.getStatusText().

Tested with Spring 3.0 and 3.1.